A news item can now have a registrationform.
The way this works is: an event now got a few additional fields and one of them is a boolean which indicates if you can register. Another field contains which data the registration form needs. There is another field in which you can specify the ultimatum for the registration to be in (not used yet).

A news item can be linked to an event so that if you press the news item and the event accepts registrations the user will be presented with the registration form.